create DTO, service & controller della classe Fattura, aggiunte anche le query alla repository

+++ EpicEnergyService/src/main/java/it/epicode/EpicEnergyService/fatture/Fattura.java
@@ -23,7 +23,7 @@ public class Fattura {
     private LocalDate data;
 
     @Column(nullable = false)
-    private Long importo;
+    private double importo;
 
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
